+++ src/app/utils/server.utils.ts
@@ -44,13 +44,6 @@ declare module 'fastify' {
 	}
 
 	/**
-	 * Server reply
-	 */
-	interface FastifyReply {
-		body: UserRecieve;
-	}
-
-	/**
 	 * Server session addition fields
 	 */
 	interface Session {
@@ -88,7 +81,7 @@ export async function registerServerPlugins(
 			brotliOptions: { params: { [constants.BROTLI_PARAM_QUALITY]: 6 } },
 		})
 		.register(fastifySecuredSession, {
-			cookieName: (18).alpha.toBase64Url,
+			cookieName: 'session',
 			cookie: { ...cookieOptions, signed: true },
 			secret,
 			key: readFileSync('securedSessionKey'),
@@ -96,7 +89,7 @@ export async function registerServerPlugins(
 		})
 		.register(fastifyCsrf, {
 			sessionKey: name,
-			cookieKey: (18 + (18).random).alpha.toBase64Url,
+			cookieKey: 'csrf',
 			cookieOpts: cookieOptions,
 			sessionPlugin: '@fastify/secure-session',
 			csrfOpts: { validity: (180).s2ms },
@@ -200,12 +193,12 @@ export class InitServerClass implements OnModuleInit {
 			middleware = new AppMiddleware(this.jwt, this.config);
 
 		adapterInstance
-			.addHook('preValidation', (req, rep) => middleware.auth(req, rep))
+			.addHook('preValidation', (req, rep, done) =>
+				middleware.auth(req, rep, done),
+			)
 			.addHook('preValidation', (req, rep) => middleware.graphQl(req, rep))
-			.addHook(
-				'preSerialization',
-				async (request, reply, payload: UserRecieve) =>
-					middleware.cookie(request, reply, payload),
+			.addHook('preSerialization', (req, rep, payload: UserRecieve, done) =>
+				middleware.cookie(req, rep, payload, done),
 			)
 			.addContentTypeParser(
 				/^multipart\/([\w-]+);?/,